@@ -4,32 +4,31 @@
   "private": true,
   "dependencies": {
     "@date-io/dayjs": "^2.16.0",
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.8.0",
-    "@mui/lab": "^5.0.0-alpha.101",
-    "@mui/material": "^5.8.0",
-    "@mui/styles": "^5.8.0",
-    "@mui/utils": "^5.10.6",
-    "@mui/x-data-grid": "^5.17.14",
-    "@mui/x-date-pickers": "^5.0.3",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.27.2",
-    "clsx": "^1.1.1",
-    "dayjs": "^1.11.2",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/lab": "^5.0.0-alpha.116",
+    "@mui/material": "^5.11.5",
+    "@mui/styles": "^5.11.2",
+    "@mui/utils": "^5.11.2",
+    "@mui/x-data-grid": "^5.17.20",
+    "@mui/x-date-pickers": "^5.0.15",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.2.3",
+    "clsx": "^1.2.1",
+    "dayjs": "^1.11.7",
     "humps": "^2.0.1",
     "lodash": "^4.17.21",
-    "notistack": "^2.0.5",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "react-hook-form": "^7.36.1",
-    "react-phone-number-input": "^3.2.2",
-    "react-router": "^6.4.0",
-    "react-router-dom": "^6.4.0",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
+    "notistack": "^2.0.8",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.42.1",
+    "react-phone-number-input": "^3.2.16",
+    "react-router": "^6.7.0",
+    "react-router-dom": "^6.7.0",
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "CHOKIDAR_USEPOLLING=true PORT=22014 FAST_REFRESH=false react-scripts start",
@@ -60,7 +59,7 @@
     ]
   },
   "devDependencies": {
-    "@trivago/prettier-plugin-sort-imports": "^3.2.0",
-    "prettier": "^2.6.0"
+    "@trivago/prettier-plugin-sort-imports": "^4.0.0",
+    "prettier": "^2.8.3"
   }
 }
