@@ -3,38 +3,37 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@sentry/browser": "^7.21.1",
-    "@sentry/tracing": "^7.21.1",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.26.1",
+    "@sentry/browser": "^7.31.1",
+    "@sentry/tracing": "^7.31.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.2.3",
     "bluebird": "^3.7.2",
-    "bootstrap": "^5.2.0",
-    "bootstrap-icons": "^1.8.1",
-    "clsx": "^1.1.1",
-    "dayjs": "^1.11.0",
+    "bootstrap": "^5.2.3",
+    "bootstrap-icons": "^1.10.3",
+    "clsx": "^1.2.1",
+    "dayjs": "^1.11.7",
     "humps": "^2.0.1",
-    "i18next": "^21.6.14",
-    "i18next-http-backend": "^1.4.0",
+    "i18next": "^22.4.9",
+    "i18next-http-backend": "^2.1.1",
     "leaflet": "^1.9.3",
     "leaflet.animatedmarker": "^1.0.0",
     "leaflet.markercluster": "^1.5.3",
     "lodash": "^4.17.21",
-    "markdown-to-jsx": "^7.1.7",
+    "markdown-to-jsx": "^7.1.8",
     "payment": "^2.4.6",
     "prop-types": "^15.8.1",
-    "qs": "^6.10.3",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.2.0",
-    "react-dom": "^17.0.2",
+    "qs": "^6.11.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.0",
+    "react-dom": "^18.2.0",
     "react-helmet-async": "^1.3.0",
-    "react-hook-form": "^7.32.0",
-    "react-phone-number-input": "^3.1.46",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "^5.0.0",
-    "sass": "^1.50.0",
-    "web-vitals": "^2.1.4"
+    "react-hook-form": "^7.42.1",
+    "react-phone-number-input": "^3.2.16",
+    "react-router-dom": "^6.7.0",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.57.1"
   },
   "scripts": {
     "analyze": "source-map-explorer ../build-webapp/static/js/**/*.js",
@@ -67,8 +66,8 @@
     ]
   },
   "devDependencies": {
-    "@trivago/prettier-plugin-sort-imports": "^3.2.0",
-    "prettier": "^2.6.0",
+    "@trivago/prettier-plugin-sort-imports": "^4.0.0",
+    "prettier": "^2.8.3",
     "qrcode": "^1.5.1",
     "source-map-explorer": "^2.5.3"
   }
